Add initial impl of Static folders
This commit is a hefty one--it adds the initial implementation of static folders. Much of the general strategy is the same as that pursued in Dynamic code. The major difference revolves around typing--since our generic folds need to type check, user's have to implement a few interfaces in order to enable support for their types in generic folds.